apps/api/pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["spektors_api*"]

[project]
name = "spektors-api"
version = "0.1.0"
description = "Spektors banking + LLM API"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pyyaml>=6.0.2",
    "httpx>=0.28.0",
    "asyncpg>=0.30.0",
    "chromadb>=0.5.0",
    "alembic>=1.13.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.9",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "pytest-asyncio>=0.24.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
asyncio_mode = "auto"
markers = [
    "integration: требует Postgres (SPEKTORS_INTEGRATION=1)",
]
